server/requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-Cors==3.0.10
google-api-core==2.10.2
google-api-python-client==2.65.0
google-auth==2.13.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.3
googleapis-common-protos==1.56.4
httplib2==0.20.4
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.2
oauthlib==3.2.2
Pillow==9.4.0
protobuf==4.21.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
slack-sdk==3.19.1
torch==1.13.1
torchvision==0.14.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.12
Werkzeug==2.2.3
zipp==3.15.0
